Moving and fixing issues
