+++ pygim-common/_pygim/_utils/_iterable.py
@@ -3,15 +3,7 @@
 This module contains internal utility functions.
 """
 
-__all__ = ("split", "flatten", "is_container", "has_instances", "is_subset")
-
-
-def has_instances(iterable, types, *, how=all):
-    return how(isinstance(it, types) for it in iterable)
-
-
-def is_subset(iterable, other):
-    return set(iterable).issubset(other)
+__all__ = ("split", "flatten", "is_container")
 
 
 def split(iterable, condition):
