--- pygim-common/pygim/_cli/cliapp.pyi
@@ -1,4 +1,0 @@
-class GimmicksCliApp:
-    def clean_up(
-        self, yes: bool, build_dirs: bool, pycache_dirs: bool, compiled_files: bool, quiet: bool, all: bool
-    ) -> None: ...
