--- pygim-common/pygim/_cli/support.pyi
@@ -1,3 +1,0 @@
-from _typeshed import Incomplete
-
-flag_opt: Incomplete
