+++ pygim-common/pyproject.toml
@@ -33,12 +33,16 @@ dependencies = [
   'coverage>=6.4.4',
   'pytest>=7.0.0',
   'typing_extensions>=4.2.0',
+  'tabulate>=0.8.10'
 ]
 
 [project.optional-dependencies]
 dev = [
+  'pytest',
   'pytest-xdist',
-  'mypy'
+  'mypy',
+  'coverage',
+  'coverage[toml]',
 ]
 
 [tool.pytest.ini_options]
